style(*): use int32 and int64 rather than int

+++ src/database/dao/SubDao.go
@@ -27,38 +27,38 @@ func SubRepository(config *config.MySqlConfig) *SubDao {
 	}
 }
 
-func (s *SubDao) QuerySubscriberUsers(uid int, page int) (users []*po.User, count int, status database.DbStatus) {
+func (s *SubDao) QuerySubscriberUsers(uid int32, page int32) (users []*po.User, count int32, status database.DbStatus) {
 	// https://gorm.io/docs/many_to_many.html
 	user := &po.User{Uid: uid}
 	if !s.userDao.Exist(uid) {
 		return nil, 0, database.DbNotFound
 	}
-	count = s.db.Model(user).Association(s.colSubscribers).Count() // 开始关联模式
+	count = int32(s.db.Model(user).Association(s.colSubscribers).Count()) // 开始关联模式
 	s.db.Limit(s.config.PageSize).Offset((page-1)*s.config.PageSize).Model(user).Related(&users, s.colSubscribers)
 	return users, count, database.DbSuccess
 }
 
-func (s *SubDao) QuerySubscribingUsers(uid int, page int) (users []*po.User, count int, status database.DbStatus) {
+func (s *SubDao) QuerySubscribingUsers(uid int32, page int32) (users []*po.User, count int32, status database.DbStatus) {
 	user := &po.User{Uid: uid}
 	if !s.userDao.Exist(uid) {
 		return nil, 0, database.DbNotFound
 	}
-	count = s.db.Model(user).Association(s.colSubscribings).Count()
+	count = int32(s.db.Model(user).Association(s.colSubscribings).Count())
 	s.db.Limit(s.config.PageSize).Offset((page-1)*s.config.PageSize).Model(user).Related(&users, s.colSubscribings)
 	return users, count, database.DbSuccess
 }
 
-func (s *SubDao) QuerySubCnt(uid int) (subscribingCnt int, subscriberCnt int, status database.DbStatus) {
+func (s *SubDao) QuerySubCnt(uid int32) (subscribingCnt int32, subscriberCnt int32, status database.DbStatus) {
 	if !s.userDao.Exist(uid) {
 		return 0, 0, database.DbNotFound
 	}
 	user := &po.User{Uid: uid}
-	subscribingCnt = s.db.Model(user).Association(s.colSubscribings).Count()
-	subscriberCnt = s.db.Model(user).Association(s.colSubscribers).Count()
+	subscribingCnt = int32(s.db.Model(user).Association(s.colSubscribings).Count())
+	subscriberCnt = int32(s.db.Model(user).Association(s.colSubscribers).Count())
 	return subscribingCnt, subscriberCnt, database.DbSuccess
 }
 
-func (s *SubDao) SubscribeUser(meUid int, toUid int) database.DbStatus {
+func (s *SubDao) SubscribeUser(meUid int32, toUid int32) database.DbStatus {
 	if !s.userDao.Exist(toUid) || !s.userDao.Exist(meUid) {
 		return database.DbNotFound
 	}
@@ -69,7 +69,7 @@ func (s *SubDao) SubscribeUser(meUid int, toUid int) database.DbStatus {
 	return database.DbSuccess
 }
 
-func (s *SubDao) UnSubscribeUser(meUid int, toUid int) database.DbStatus {
+func (s *SubDao) UnSubscribeUser(meUid int32, toUid int32) database.DbStatus {
 	if !s.userDao.Exist(toUid) || !s.userDao.Exist(meUid) {
 		return database.DbNotFound
 	}
